backend/requirements.txt
uWSGI
paddlepaddle==2.2.2
Pillow
django
urllib3
virtualenv
asgiref
django-rest-framework
PyMySQL
mysqlclient
redis
